requirements.txt
APScheduler==3.6.3
Babel==2.9.1
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coverage==5.5
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==1.1.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.4.0
Flask-Log-Request-ID==0.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.15.1
flipflop==1.0
greenlet==1.1.0
guess-language==0.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pbr==5.6.0
pyTelegramBotAPI==3.7.9
python-telegram-bot==13.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.25.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.18
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
Tempita==0.5.2
tornado==6.1
tzlocal==2.1
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
visitor==0.1.3
Werkzeug==2.0.1
Whoosh==2.7.4
WTForms==2.3.3
xlwt==1.3.0
